fix: surface BLE connect failures, and stop C++ enums logging as garbage (#375)
* fix: surface BLE connect failures, and stop enums logging as garbage

A user report of "Android can't connect on app v3.3.0" root-caused to a bond
mismatch, not an app regression: the board's settings partition had been wiped
(HIL destructive tier / appcfg erase) while the phone kept its bond, so every
connect died in ~370 ms with BT_SECURITY_ERR_PIN_OR_KEY_MISSING before discovery
started. Reproduced live on a Pixel 9 Pro with the v3.3.0 release build; clearing
the stale bond fixed it (verified CONNECTED / L4 / ATT MTU 498).

Two defects made that far more expensive to diagnose than it should have been:

app: a failed connect was completely silent. connect() resolves false and the
press handler simply doesn't navigate, so every failure looked like a dead
button. Device rows now show a tappable failure indicator whose alert carries the
underlying reason plus the forget-and-re-pair recovery. Deliberately NOT
classified by error code: describeWriteError only asserts androidErrorCode 252
because that was hardware-measured (issue #92), and the connect-side ble-plx shape
has not been captured, so the reason is surfaced verbatim and the hint is phrased
as guidance rather than a diagnosis.

fw: security levels printed as garbage - "Failed to reach required security level
59609092, got 59609089" instead of 4 and 1. Under CONFIG_LOG_MODE_DEFERRED
arguments are packed, not passed as varargs, so no integer promotion happens;
cbprintf_cxx.h has promoting overloads for char/short/etc but none for enums, so
an enum takes the generic template, which reserves MAX(sizeof(T), sizeof(int))
bytes and then copies 4 bytes out of a 1-byte object - the true value survives in
the low byte, the upper 24 bits are adjacent stack. bt_security_t is
`typedef enum __packed`. This affects EVERY enum logged from C++ here, so the rule
is documented in fw/CLAUDE.md rather than only fixed at these six call sites.
Hardware-verified after flashing: "Peer security level changed to 4".

scripts: re-pair.py hardcoded the dev package, so pointing it at a release-build
bug silently drove the wrong app and reported "board not listed" - which blames
the board rather than the app. Adds --package/--release (default unchanged), a
fail-fast install check, and a startup line naming the package driven.

Verified: proto0 builds clean; firmware flashed and confirmed on hardware;
app tsc/lint clean, jest 546/546 (+9 new tests); re-pair --release re-paired and
verified CONNECTED / L4 / MTU 498.


* fix: match cast types to format specifiers; document the never-exits launch trap

Self-review + hardware verification follow-ups on the previous commit:

- fw: the security-level logs use %u, so cast to (unsigned) rather than (int).
  Behaviourally identical for the 0-4 values these carry, but a reviewer would
  rightly flag an (int) argument under a %u specifier.

- app: two import orderings restored to alphabetical (icon-symbol after
  app-button; ble-errors before ble-foreground-service). Lint does not enforce
  ordering, so these were invisible to the gate.

- app/CLAUDE.md: record two traps hit while verifying this branch on hardware.
  `launch-app.sh` never exits (it exec's into Metro, which holds the foreground
  for the session), so waiting for the task to "complete" blocks forever while
  the app has in fact been installed and running the whole time — the launch is
  done when the log says BUILD SUCCESSFUL / Installing / Opening / Bundled, not
  when the task ends. Also notes that redirecting its output to a file takes the
  app's console.log out of the tracked task stream, and that `adb logcat -s
  ReactNativeJS` is where those lines actually are.


* fix: extend the cast rule to bool, and stop re-pair's filter defeating --package

Addresses PR #375 review (all three findings confirmed against the SDK/build
before acting):

- fw: `pairing_complete()`'s `bonded` is a `bool`, and cbprintf_cxx.h has no
  `bool` overload either — a bool is an EXACT match for the generic template
  (beating every conversion-requiring candidate), so it took the identical
  4-bytes-out-of-a-1-byte-object path and a successful pairing logged
  "bonded: 59609089" instead of "bonded: 1". Now cast.

- fw/CLAUDE.md: the "enum class is safe by construction" line was WRONG and would
  have let the bug straight back in. Zephyr's arg checker is only `__printf_like`,
  so a mismatch is a -Wformat warning, and this build sets no -Werror (verified:
  neither fw/prj.conf nor fw/CMakeLists.txt) — a scoped enum under %d compiles
  clean and ships the same garbage, while an unscoped packed enum warns not at
  all. Rule widened to every enum, scoped or unscoped, and every bool, with the
  warning explicitly demoted from gate to bonus.

- scripts: re-pair.py's install check filtered on a hardcoded "rgbsunglasses",
  which defeats --package for any renamed/forked/staging id — an installed
  package would be reported as missing, the same misleading-failure class the
  check was added to remove. Filters on PKG itself now (keeping the exact
  "package:<name>" comparison, since `pm list packages` substring-matches), and
  derives the failure-message candidate list from PKG's own vendor prefix so a
  renamed package still gets useful suggestions instead of an empty list.

Verified: proto0 builds clean; twister 660/660 cases across 63/63 configs; app
jest 546/546, tsc clean, lint clean; re-pair.py compiles.

Note on size: appcore moved 85.08% -> 85.23% FLASH and 88.15% -> 88.47% RAM, but
both regions moved by EXACTLY +1408 B, which is the kobject_data signature (that
section is counted in both, and its gperf hash resizes on any layout shift) — not
the cost of one cast. See the kobject_data note in the root CLAUDE.md.


---------

+++ fw/src/bluetooth.cpp
@@ -500,10 +500,21 @@ static void security_changed(struct bt_conn *conn, bt_security_t level, enum bt_
 
     bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
 
+    // (int) casts are REQUIRED, not cosmetic: bt_security_t is `typedef enum __packed`
+    // (1 byte) and this is a C++ TU, so Zephyr's log-argument packing takes the generic
+    // template in cbprintf_cxx.h - which reserves MAX(sizeof(T), sizeof(int)) = 4 bytes
+    // but then memcpy's 4 bytes out of the 1-byte object, pulling 3 bytes of adjacent
+    // stack into the upper bits. Without the cast this printed "level 66701313 err
+    // 66701314" (0x03F9C801 / 0x03F9C802) instead of "level 1 err 2" - the true value
+    // survives in the low byte only. The explicit char/short overloads in that header
+    // promote via `int tmp = arg + 0;`; enums have no such overload, so the promotion
+    // has to happen here. Applies to EVERY enum logged from C++ in this codebase - see
+    // fw/CLAUDE.md, "Logging an enum from C++ prints garbage without an (int) cast".
+    // Cast type matches each specifier: (unsigned) for %u, (int) for %d.
     if (!err) {
-        LOG_INF("Security changed: %s level %u", addr, level);
+        LOG_INF("Security changed: %s level %u", addr, (unsigned)level);
     } else {
-        LOG_ERR("Security failed: %s level %u err %d", addr, level, err);
+        LOG_ERR("Security failed: %s level %u err %d", addr, (unsigned)level, (int)err);
     }
 
     BtThreadCommand cmd;
@@ -565,7 +576,12 @@ static void pairing_complete(struct bt_conn *conn, bool bonded) {
 
     bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
 
-    LOG_INF("Pairing completed: %s, bonded: %d", addr, bonded);
+    // (int) for the same reason as the enum sites below: cbprintf_cxx.h has explicit
+    // promoting overloads for char/short/etc but none for `bool` either, so a bool is an
+    // EXACT match for the generic template (beating every conversion-requiring candidate)
+    // and takes the identical 4-bytes-out-of-a-1-byte-object path. Uncast, a successful
+    // pairing logged "bonded: 59609089" instead of "bonded: 1".
+    LOG_INF("Pairing completed: %s, bonded: %d", addr, (int)bonded);
 }
 
 static void pairing_failed(struct bt_conn *conn, enum bt_security_err reason) {
@@ -573,7 +589,9 @@ static void pairing_failed(struct bt_conn *conn, enum bt_security_err reason) {
 
     bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
 
-    LOG_INF("Pairing failed conn: %s, reason %d", addr, reason);
+    // (int): enum bt_security_err is 1 byte under -fshort-enums - see the cast rationale
+    // in security_changed() above.
+    LOG_INF("Pairing failed conn: %s, reason %d", addr, (int)reason);
 }
 
 static void bond_deleted(uint8_t id, const bt_addr_le_t *peer) {
@@ -813,7 +831,8 @@ void bt_state_connecting_handle_command(BtThreadContext *ctx, const BtThreadComm
 
         case BtThreadEvent::SECURITY_CHANGED:
             // What level of security did the peer upgrade to?
-            LOG_INF("Peer security level changed to %d", cmd->level);
+            // (int): bt_security_t is a 1-byte packed enum - see security_changed().
+            LOG_INF("Peer security level changed to %d", (int)cmd->level);
 
             if (cmd->level == REQUIRED_BT_SECURITY_LEVEL) {
                 LOG_DBG("Required security level achieved");
@@ -838,8 +857,11 @@ void bt_state_connecting_handle_command(BtThreadContext *ctx, const BtThreadComm
 
                 bt_state_change_to(ctx, BtThreadState::CONNECTED);
             } else {
+                // (int) on BOTH: REQUIRED_BT_SECURITY_LEVEL is BT_SECURITY_L4, which is
+                // itself a bt_security_t constant - so even the compile-time literal
+                // printed as 0x038D9004 instead of 4. See security_changed().
                 LOG_ERR("Failed to reach required security level %d, got %d instead",
-                        REQUIRED_BT_SECURITY_LEVEL, cmd->level);
+                        (int)REQUIRED_BT_SECURITY_LEVEL, (int)cmd->level);
 
                 // Disconnect from the peer (re-enabled for issue #232). With
                 // CONFIG_BT_SMP_SC_ONLY a below-L4 pairing can no longer *succeed*,
@@ -901,14 +923,15 @@ void bt_state_connected_handle_command(BtThreadContext *ctx, const BtThreadComma
 
         case BtThreadEvent::SECURITY_CHANGED:
             // What level of security did the peer upgrade to?
-            LOG_INF("Peer security level changed to %d", cmd->level);
+            // (int): bt_security_t is a 1-byte packed enum - see security_changed().
+            LOG_INF("Peer security level changed to %d", (int)cmd->level);
 
             if (cmd->level == REQUIRED_BT_SECURITY_LEVEL) {
                 LOG_DBG("Required security level achieved");
                 bt_state_change_to(ctx, BtThreadState::CONNECTED);
             } else {
-                LOG_WRN("Reached security %d, must reach %d to connect", cmd->level,
-                        REQUIRED_BT_SECURITY_LEVEL);
+                LOG_WRN("Reached security %d, must reach %d to connect", (int)cmd->level,
+                        (int)REQUIRED_BT_SECURITY_LEVEL);
                 // TODO: need to implement timeout to disconnect if we don't reach "CONNECTED" state
                 // within ~60s
             }
